django_ui/requirements.txt
django-filter~=22.1.0
djangorestframework-gis~=1.0.0
djangorestframework~=3.14.0
django~=4.2.2
sqlparse==0.4.4